+++ backend/src/main/java/com/proyecto/erpventas/infrastructure/repository/factura/SpringFacturaJpaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface SpringFacturaJpaRepository extends JpaRepository<Factura, Long> {
+public interface SpringFacturaJpaRepository extends JpaRepository<Factura, Integer> {
     boolean existsByNumeroFactura(String numeroFactura);
     Optional<Factura> findByNumeroFactura(String numeroFactura);
     List<Factura> findAllByActivoTrueOrderByFechaEmisionDesc();
